unit/requirements.txt
anyio
certifi
charset-normalizer
click
dnspython
email-validator
fastapi
h11
httptools
idna
itsdangerous
Jinja2
MarkupSafe
orjson
pip
pydantic
pymongo
python-dotenv
python-multipart
PyYAML
requests
setuptools
six
sniffio
starlette
typing_extensions
ujson
urllib3
uvicorn
uvloop
websockets